feat: same as https://github.com/FreerGit/decay/commit/289cbca17fa3352dcf557a7375658a1dc1fcbdfa but with wrapper for proper parsing

+++ src/exchange/client.rs
@@ -47,7 +47,7 @@ pub struct OrderResult {
     side: Side,
     order_type: OrderType,
     price: i32,
-    qty: i32,
+    qty: Decimal,
     // @TODO Type order status?
     order_status: String,
 }
